fixed test code
