+++ llama-index-integrations/extractors/llama-index-extractors-document-context/llama_index/extractors/document_context/__init__.py
@@ -1,0 +1,3 @@
+from llama_index.extractors.document_context.base import DocumentContextExtractor
+
+__all__ = ["DocumentContextExtractor"]
